Улучшения

+++ gateway/src/main/java/ru/practicum/shareit/booking/BookingClient.java
@@ -30,7 +30,7 @@ public class BookingClient extends BaseClient {
         );
     }
 
-    public ResponseEntity<Object> getBookings(long userId, BookingState state, long from, long size) {
+    public ResponseEntity<Object> getBookings(long userId, BookingState state, Integer from, Integer size) {
         Map<String, Object> parameters = Map.of(
                 "state", state.name(),
                 "from", from,
@@ -39,7 +39,7 @@ public class BookingClient extends BaseClient {
         return get("?state={state}&from={from}&size={size}", userId, parameters);
     }
 
-    public ResponseEntity<Object> getBooking(long userId, long bookingId) {
+    public ResponseEntity<Object> getBooking(long userId, Long bookingId) {
         return get("/" + bookingId, userId);
     }
 
@@ -47,7 +47,7 @@ public class BookingClient extends BaseClient {
         return post("", userId, requestDto);
     }
 
-    public ResponseEntity<Object> approveBooking(long bookingId, long userId, boolean approved) {
+    public ResponseEntity<Object> approveBooking(long bookingId, Long userId, boolean approved) {
         log.info("Approving booking with id: {} by user with id: {}", bookingId, userId);
         Map<String, Object> parameters = Map.of("approved", approved);
         return patch("/" + bookingId + "?approved={approved}", userId, parameters, null);
